+++ lms/lms/doctype/loan_application/loan_application.py
@@ -45,21 +45,13 @@ class LoanApplication(Document):
             "borrower_name": user_kyc.investor_name,
             "borrower_address": user_kyc.address,
             "sanctioned_amount": lms.validate_rupees(
-                lms.round_down_amount_to_nearest_thousand(
-                    (self.total_collateral_value + loan.total_collateral_value)
-                    * self.allowable_ltv
-                    / 100
-                )
-            )
-            if self.loan and not self.loan_margin_shortfall
-            else int(self.drawing_power),
+                self.increased_sanctioned_limit
+                if self.loan and not self.loan_margin_shortfall
+                else self.drawing_power
+            ),
             "sanctioned_amount_in_words": lms.number_to_word(
                 lms.validate_rupees(
-                    lms.round_down_amount_to_nearest_thousand(
-                        (self.total_collateral_value + loan.total_collateral_value)
-                        * self.allowable_ltv
-                        / 100
-                    )
+                    self.increased_sanctioned_limit
                     if self.loan and not self.loan_margin_shortfall
                     else self.drawing_power,
                 )
@@ -161,10 +153,15 @@ class LoanApplication(Document):
         }
 
     def before_save(self):
+        lender = self.get_lender()
+        self.minimum_sanctioned_limit = lender.minimum_sanctioned_limit
+        self.maximum_sanctioned_limit = lender.maximum_sanctioned_limit
+
         if (
             self.status == "Approved"
             and not self.lender_esigned_document
             and not self.loan_margin_shortfall
+            and not self.application_type == "Pledge More"
         ):
             frappe.throw("Please upload Lender Esigned Document")
         elif self.status == "Approved":
@@ -211,12 +208,39 @@ class LoanApplication(Document):
 
             # TODO : manage loan application and its item's as per lender approval
             self.total_collateral_value = round(total_collateral_value, 2)
-            self.drawing_power = round(
+            drawing_power = round(
                 lms.round_down_amount_to_nearest_thousand(
                     (self.allowable_ltv / 100) * self.total_collateral_value
                 ),
                 2,
             )
+            self.drawing_power = (
+                drawing_power
+                if drawing_power < self.maximum_sanctioned_limit
+                else self.maximum_sanctioned_limit
+            )
+
+            # TODO : if increase loan drawing power is less than 10k the loan application wont be proceed
+            loan_total_collateral_value = 0
+            if self.loan:
+                loan_total_collateral_value = frappe.get_doc(
+                    "Loan", self.loan
+                ).total_collateral_value
+
+            # Use increased sanctioned limit field for this validation
+            drawing_power = round(
+                lms.round_down_amount_to_nearest_thousand(
+                    (self.total_collateral_value + loan_total_collateral_value)
+                    * (self.allowable_ltv / 100)
+                ),
+                2,
+            )
+
+            if self.application_type in ["New Loan", "Increase Loan"]:
+                if drawing_power < self.minimum_sanctioned_limit:
+                    frappe.throw(
+                        "Sorry! This Loan Application can not be Approved as its Drawing power is less than Minimum Sanctioned Limit."
+                    )
 
         if self.status == "Pledge executed":
             total_collateral_value = 0
@@ -228,12 +252,18 @@ class LoanApplication(Document):
                     ):
                         total_collateral_value += i.amount
                         self.total_collateral_value = round(total_collateral_value, 2)
-                        self.drawing_power = round(
+                        drawing_power = round(
                             lms.round_down_amount_to_nearest_thousand(
                                 (self.allowable_ltv / 100) * self.total_collateral_value
                             ),
                             2,
                         )
+                        self.drawing_power = (
+                            drawing_power
+                            if drawing_power < self.maximum_sanctioned_limit
+                            else self.maximum_sanctioned_limit
+                        )
+
                     # elif (
                     #     i.lender_approval_status == "Rejected"
                     #     or i.lender_approval_status == "Pledge Failure"
@@ -579,13 +609,14 @@ class LoanApplication(Document):
 
         # apply processing fees on new sanctioned limit(loan application drawing power)
         # apply renewal charges on existing loan sanctioned limit
-        if not self.loan_margin_shortfall:
+        if self.application_type == "Increase Loan":
             self.apply_renewal_charges(loan)
-
-        loan.reload()
-        loan.update_items()
-        loan.fill_items()
-        # loan.check_for_shortfall()
+            loan.reload()
+            loan.check_for_shortfall()
+        else:
+            loan.reload()
+            loan.update_items()
+            loan.fill_items()
 
         # for item in self.items:
         #     if item.lender_approval_status == "Approved":
@@ -605,29 +636,36 @@ class LoanApplication(Document):
         loan.drawing_power = (loan.allowable_ltv / 100) * loan.total_collateral_value
         # loan.drawing_power += self.drawing_power
 
-        if not self.loan_margin_shortfall:
-            loan.drawing_power = round(
-                lms.round_down_amount_to_nearest_thousand(loan.drawing_power), 2
+        if self.application_type == "Increase Loan":
+            # drawing_power = round(
+            #     lms.round_down_amount_to_nearest_thousand(loan.drawing_power), 2
+            # )
+            # loan.sanctioned_limit = loan.drawing_power
+            loan.sanctioned_limit = loan.drawing_power = (
+                self.maximum_sanctioned_limit
+                if self.increased_sanctioned_limit > self.maximum_sanctioned_limit
+                else self.increased_sanctioned_limit
             )
-            loan.sanctioned_limit = loan.drawing_power
+
             # TODO : manage expiry date
             loan.expiry_date = self.expiry_date
             loan.save(ignore_permissions=True)
 
-        if self.loan_margin_shortfall:
+        if self.application_type in ["Margin Shortfall", "Pledge More"]:
             if loan.drawing_power > loan.sanctioned_limit:
                 loan.drawing_power = loan.sanctioned_limit
 
             loan.save(ignore_permissions=True)
-            loan_margin_shortfall = frappe.get_doc(
-                "Loan Margin Shortfall", self.loan_margin_shortfall
-            )
-            loan_margin_shortfall.fill_items()
-            # if not loan_margin_shortfall.margin_shortfall_action:
-            if loan_margin_shortfall.shortfall_percentage == 0:
-                loan_margin_shortfall.status = "Pledged Securities"
-                loan_margin_shortfall.action_time = frappe.utils.now_datetime()
-            loan_margin_shortfall.save(ignore_permissions=True)
+            if self.loan_margin_shortfall:
+                loan_margin_shortfall = frappe.get_doc(
+                    "Loan Margin Shortfall", self.loan_margin_shortfall
+                )
+                loan_margin_shortfall.fill_items()
+                # if not loan_margin_shortfall.margin_shortfall_action:
+                if loan_margin_shortfall.shortfall_percentage == 0:
+                    loan_margin_shortfall.status = "Pledged Securities"
+                    loan_margin_shortfall.action_time = frappe.utils.now_datetime()
+                loan_margin_shortfall.save(ignore_permissions=True)
 
         return loan
 
@@ -647,11 +685,7 @@ class LoanApplication(Document):
         # -> no processing fee
         # -> renewal on new sanctioned
         # new sanctioned limit = lms.round_down_amount_to_nearest_thousand((new total coll + old total coll) / 2)
-        new_sanctioned_limit = lms.round_down_amount_to_nearest_thousand(
-            (self.total_collateral_value + loan.total_collateral_value)
-            * self.allowable_ltv
-            / 100
-        )
+        new_sanctioned_limit = self.increased_sanctioned_limit
 
         renewal_sanctioned_limit, processing_sanctioned_limit = (
             (loan.sanctioned_limit, (new_sanctioned_limit - loan.sanctioned_limit))