+++ lms/user.py
@@ -1626,6 +1626,7 @@ def dashboard(**kwargs):
                     )
                     if loan_application_doc.loan
                     and not loan_application_doc.loan_margin_shortfall
+                    and not loan_application_doc.application_type == "Pledge More"
                     else "Congratulations! Your application is being considered favourably by our lending partner and finally accepted at Rs. {current_total_collateral_value} against the request value of Rs. {requested_total_collateral_value}. Accordingly the final Sanctioned Limit is Rs. {drawing_power}. Please e-sign the loan agreement to avail the loan now.".format(
                         current_total_collateral_value=frappe.utils.fmt_money(
                             loan_application_doc.total_collateral_value
@@ -1641,27 +1642,22 @@ def dashboard(**kwargs):
                 if (
                     loan_application_doc.loan
                     and not loan_application_doc.loan_margin_shortfall
+                    and not loan_application_doc.application_type == "Pledge More"
                 ):
                     loan = frappe.get_doc("Loan", loan_application_doc.loan)
 
                     increase_loan_mess = dict(
                         existing_limit=loan.sanctioned_limit,
                         existing_collateral_value=loan.total_collateral_value,
-                        new_limit=(
-                            lms.round_down_amount_to_nearest_thousand(
-                                (
-                                    loan_application_doc.total_collateral_value
-                                    + loan.total_collateral_value
-                                )
-                                * loan_application_doc.allowable_ltv
-                                / 100
-                            )
-                        ),
+                        new_limit=loan_application_doc.increased_sanctioned_limit,
                         new_collateral_value=loan_application_doc.total_collateral_value
                         + loan.total_collateral_value,
                     )
 
-                if not loan_application_doc.loan_margin_shortfall:
+                if (
+                    not loan_application_doc.loan_margin_shortfall
+                    and not loan_application_doc.application_type == "Pledge More"
+                ):
                     la_pending_esigns.append(
                         {
                             "loan_application": loan_application_doc,