+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 - [Feat: Open graph code added for (non www)spark.loans](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/406)
 - [Feat: contact us email content formatted](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/408)
 - [Feat: contact us email content changes for new line](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/409)
+- [Feat: Min max sanctioned limit](https://gitlab.com/atri-tech/atri-maintainers/choicelms/-/merge_requests/414)
 
 ### Fixed